Add new files for project 'netty-proto'
Create a new client-server model project named 'netty-proto'. The project includes three modules: server, client, and common module. The common module contains common files used by both the client and server. The server and client modules are the main application modules to start a server or a client. This project is using 'protostuff' for serialization and deserialization between the client and server, and the 'netty' framework for network communication.