Add platform-specific channel selection in Netty setup
This commit introduces a new utility class, NetworkUtil, that figures out the underlying OS and determines whether EPOLL is available for use. Based on this information, either EpollSocketChannel or NioSocketChannel, and EpollServerSocketChannel or NioServerSocketChannel are used in the NettyClient and NodesServer classes, respectively. This allows to utilize the high performance of EPOLL on Linux platforms while maintaining compatibility with other platforms.